Moving Scalars to an internal class & Updating tests to use the internal types
