+++ src/test/groovy/graphql/scalars/java/ScalarsCharTest.groovy
@@ -1,4 +1,4 @@
-package graphql.scalars
+package graphql.scalars.java
 
 import graphql.Scalars
 import graphql.language.IntValue
@@ -14,7 +14,7 @@ class ScalarsCharTest extends Specification {
     @Unroll
     def "Char parse literal #literal.value as #result"() {
         expect:
-        Scalars.GraphQLChar.getCoercing().parseLiteral(literal) == result
+        JavaPrimitives.GraphQLChar.getCoercing().parseLiteral(literal) == result
 
         where:
         literal              | result
@@ -26,7 +26,7 @@ class ScalarsCharTest extends Specification {
     @Unroll
     def "Short returns null for invalid #literal"() {
         when:
-        Scalars.GraphQLChar.getCoercing().parseLiteral(literal)
+        JavaPrimitives.GraphQLChar.getCoercing().parseLiteral(literal)
         then:
         thrown(CoercingParseLiteralException)
 
@@ -39,8 +39,8 @@ class ScalarsCharTest extends Specification {
     @Unroll
     def "Short serialize #value into #result (#result.class)"() {
         expect:
-        Scalars.GraphQLChar.getCoercing().serialize(value) == result
-        Scalars.GraphQLChar.getCoercing().parseValue(value) == result
+        JavaPrimitives.GraphQLChar.getCoercing().serialize(value) == result
+        JavaPrimitives.GraphQLChar.getCoercing().parseValue(value) == result
 
         where:
         value | result
@@ -51,7 +51,7 @@ class ScalarsCharTest extends Specification {
     @Unroll
     def "serialize throws exception for invalid input #value"() {
         when:
-        Scalars.GraphQLChar.getCoercing().serialize(value)
+        JavaPrimitives.GraphQLChar.getCoercing().serialize(value)
         then:
         thrown(CoercingSerializeException)
 
@@ -66,7 +66,7 @@ class ScalarsCharTest extends Specification {
     @Unroll
     def "parseValue throws exception for invalid input #value"() {
         when:
-        Scalars.GraphQLChar.getCoercing().parseValue(value)
+        JavaPrimitives.GraphQLChar.getCoercing().parseValue(value)
         then:
         thrown(CoercingParseValueException)
